extract HiveHepExtractRelNodeRule
