+++ ql/src/java/org/apache/hadoop/hive/ql/optimizer/calcite/rules/HiveFieldTrimmerRule.java
@@ -19,9 +19,6 @@ package org.apache.hadoop.hive.ql.optimizer.calcite.rules;
 
 import org.apache.calcite.plan.RelOptRule;
 import org.apache.calcite.plan.RelOptRuleCall;
-import org.apache.calcite.plan.hep.HepPlanner;
-import org.apache.calcite.plan.hep.HepProgram;
-import org.apache.calcite.plan.hep.HepProgramBuilder;
 import org.apache.calcite.plan.hep.HepRelVertex;
 import org.apache.calcite.rel.RelNode;
 import org.apache.hadoop.hive.ql.optimizer.calcite.HiveRelFactories;
@@ -31,10 +28,6 @@ import org.apache.hadoop.hive.ql.optimizer.calcite.HiveRelFactories;
  * Rule that triggers the field trimmer on the root of a plan.
  */
 public class HiveFieldTrimmerRule  extends RelOptRule {
-
-  private static final HepProgram PROGRAM = new HepProgramBuilder()
-      .addRuleInstance(HiveHepExtractRelNodeRule.INSTANCE)
-      .build();
 
   private final boolean fetchStats;
   private boolean triggered;
@@ -64,9 +57,7 @@ public class HiveFieldTrimmerRule  extends RelOptRule {
       return;
     }
     // The node is the root, release the kraken!
-    final HepPlanner tmpPlanner = new HepPlanner(PROGRAM);
-    tmpPlanner.setRoot(node);
-    node = tmpPlanner.findBestExp();
+    node = HiveHepExtractRelNodeRule.execute(node);
     call.transformTo(trim(call, node));
     triggered = true;
   }
@@ -74,25 +65,4 @@ public class HiveFieldTrimmerRule  extends RelOptRule {
   protected RelNode trim(RelOptRuleCall call, RelNode node) {
     return HiveRelFieldTrimmer.get(fetchStats).trim(call.builder(), node);
   }
-
-  /**
-   * The goal of this rule is to extract the RelNode from the
-   * HepRelVertex node so the trimmer can be applied correctly.
-   */
-  private static class HiveHepExtractRelNodeRule extends RelOptRule {
-
-    private static final HiveHepExtractRelNodeRule INSTANCE =
-        new HiveHepExtractRelNodeRule();
-
-    private HiveHepExtractRelNodeRule() {
-      super(operand(HepRelVertex.class, any()));
-    }
-
-    @Override
-    public void onMatch(RelOptRuleCall call) {
-      final HepRelVertex rel = call.rel(0);
-      call.transformTo(rel.getCurrentRel());
-    }
-  }
-
 }
